requirements.txt
fastapi
uvicorn
psutil
docker
PyJWT
python-dotenv
pydantic
